fix more lint
