+++ packages/studio-base/src/panels/Image/hooks/synchronizedAddMessages.ts
@@ -9,7 +9,7 @@ import { MessageEvent } from "@foxglove/studio";
 import { normalizeCameraInfo } from "./normalizeCameraInfo";
 import { normalizeAnnotations } from "../lib/normalizeAnnotations";
 import { normalizeImageMessage } from "../lib/normalizeMessage";
-import { Annotation, ImagePanelState } from "../types";
+import type { Annotation, ImagePanelState, SynchronizationItem } from "../types";
 
 export function synchronizedAddMessages(
   state: Pick<ImagePanelState, "imageTopic" | "cameraInfoTopic" | "annotationTopics" | "tree">,
