+++ packages/studio-base/src/panels/Plot/types.ts
@@ -2,15 +2,9 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import type { BasePlotPath, PlotPath } from "@foxglove/studio-base/panels/Plot/internalTypes";
 import { PANEL_TITLE_CONFIG_KEY } from "@foxglove/studio-base/util/layout";
 
-// X-axis values:
-export type PlotXAxisVal =
-  | "timestamp" // Message playback time. Preloaded.
-  | "index" // Message-path value index. One "current" message at playback time.
-  | "custom" // Message path data. Preloaded.
-  | "currentCustom"; // Message path data. One "current" message at playback time.
+import type { BasePlotPath, PlotPath, PlotXAxisVal } from "./internalTypes";
 
 /**
  * Coalesces null, undefined and empty string to undefined.
@@ -31,6 +25,7 @@ type DeprecatedPlotConfig = {
   showSidebar?: boolean;
   sidebarWidth?: number;
 };
+
 export type PlotConfig = DeprecatedPlotConfig & {
   /** @deprecated Replaced by global panel rename functionality https://github.com/foxglove/studio/pull/5225 */
   title?: string;
